+++ changelog.md
@@ -2,9 +2,15 @@
 Notable changes to the `cc.isr.vi` solution are documented in this file using the 
 [Keep a Changelog] style. The dates specified are in coordinated universal time (UTC).
 
-[8.1.9412]: https://www.github.com/atecoder/dn.vi.ivi
+[8.1.9414]: https://www.github.com/atecoder/dn.vi.ivi
 
 Current IVI Visa Compatibility: 8.0.1.9391
+
+## [8.1.9414] - 2025-10-10
+- Device.Tsp
+  - Load: Add load method with conversion to byte code.
+  - Export:
+    - Pass the export folder name as an argument to the export method.
 
 ## [8.1.9412] - 2025-10-08
 - Visa